Finish re-organization of controllers: - Add Character directory, make all necessary changes to routes - Create controllers for each page of the form, make all necessary changes to routes
