+++ resources/views/characters/Form/alignQuest.blade.php
@@ -4,7 +4,7 @@
 <div class="container text-center">
     <h1>What is your Character's Alignment?</h1>
     
-    {!! Form::open(['action' => 'CharactersController@alignQuest', 'method' => 'POST', 'enctype' => 'multipart/form-data']) !!}
+    {!! Form::open(['action' => 'Character\AlignmentController@alignQuest', 'method' => 'POST', 'enctype' => 'multipart/form-data']) !!}
         {{ csrf_field() }}
         <div class="form-group">
             <h3>{{Form::label('What is your Character\'s Alignment?')}}</h3>
