+++ resources/views/characters/Form/raceQuest.blade.php
@@ -4,7 +4,7 @@
 <div class="container text-center">
     <h1>What is your Character's Race?</h1>
 
-    {!! Form::open(['action' => 'CharactersController@raceQuest', 'method' => 'POST', 'enctype' => 'multipart/form-data']) !!}
+    {!! Form::open(['action' => 'Character\RaceController@raceQuest', 'method' => 'POST', 'enctype' => 'multipart/form-data']) !!}
         <div class="form-group">
             <h3>{{Form::label('What Race is your Character?')}}</h3>
             <p>
