+++ resources/views/dashboard.blade.php
@@ -40,7 +40,7 @@
                                     <tr>
                                         <th><a href="/characters/{{$character->id}}">{{$character->character_name}}</a></th>
                                     <th><a href="#" class="btn btn-primary">Edit</a></th>
-                                        <th>{!!Form::open(['action' => ['CharactersController@destroy', $character->id], 'method' => 'POST'])!!}
+                                        <th>{!!Form::open(['action' => ['Character\CharactersController@destroy', $character->id], 'method' => 'POST'])!!}
                                     {{Form::hidden('_method', 'DELETE')}}
                                     {{Form::submit('Delete?', ['class' => 'ham'])}}
                                 {!!Form::close()!!}</th>
